+++ module/Shop/Service/ProductManager.php
@@ -575,6 +575,18 @@ final class ProductManager extends AbstractManager implements ProductManagerInte
 	}
 
 	/**
+	 * Fetches all published product entities with maximal view counts
+	 * 
+	 * @param integer $limit Fetching limit
+	 * @param string $categoryId Optionally can be filtered by category id
+	 * @return array
+	 */
+	public function fetchAllPublishedWithMaxViewCount($limit, $categoryId = null)
+	{
+		return $this->prepareResults($this->productMapper->fetchAllPublishedWithMaxViewCount($limit, $categoryId));
+	}
+
+	/**
 	 * Fetches all published product entities associated with given category id
 	 * 
 	 * @param string $categoryId