+++ module/Shop/Service/SiteService.php
@@ -52,6 +52,18 @@ final class SiteService implements SiteServiceInterface
 	}
 
 	/**
+	 * Returns an array of entities with products that have maximal view count
+	 * 
+	 * @param integer $limit
+	 * @param string $categoryId Optionally can be filtered by category id
+	 * @return array
+	 */
+	public function getProductsWithMaxViewCount($limit, $categoryId = null)
+	{
+		return $this->productManager->getProductsWithMaxViewCount($limit, $categoryId);
+	}
+
+	/**
 	 * Returns an array of entities of recent products
 	 * 
 	 * @param string $id Current product id to be excluded